// src/types/DependencyRegistry.ts
export enum DependencyRegistry {
	Application = 'Application',
	ILoggerService = 'ILoggerService',
	IAuthorizationController = 'IAuthorizationController',
	IUserService = 'IUserService',
	IMongoService = 'IMongoService',
	IUserRepository = 'IUserRepository',
	IExceptionFilter = 'IExceptionFilter',
}